crud products, Transactions, detail + history Transactions

+++ database/migrations/2023_07_11_031354_products.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name', 120);
-            $table->integer('price')->default(0);
+            $table->bigInteger('price')->default(0);
             $table->integer('quantity')->default(0);
             $table->timestamps();
             $table->softDeletes($column = 'deleted_at', $precision = 0);
